Use dedicated function to determine workspace size for alogorithm.
The workspace size used for the fastest may not be the workspace
size required when running the same algorithm, which may happen if
the same algorithm was profiled with different math type (FP32 and
FP16) requiring different workspace sizes. This fix uses the simple
approach of getting the max required workspace size from a cudnn
function (another approach would be to coalesce the profile
results).